Update azure scripts - Added Kafka on Storm support, VNet support, logging improvement

+++ scripts/azure/HDInsight/DeleteCluster.ps1
@@ -6,5 +6,24 @@ Param(
     [String]$ClusterName                     # required    needs to be alphanumeric or "-"
     )
 
+###########################################################
+# Start - Initialization - Invocation, Logging etc
+###########################################################
+$VerbosePreference = "SilentlyContinue"
+$ErrorActionPreference = "Stop"
+
+$scriptPath = $MyInvocation.MyCommand.Path
+$scriptDir = Split-Path $scriptPath
+
+& "$scriptDir\..\..\init.ps1"
+if(-not $?)
+{
+    throw "Initialization failure."
+    exit -9999
+}
+###########################################################
+# End - Initialization - Invocation, Logging etc
+###########################################################
+
 Write-InfoLog "Deleting Azure HDInsight cluster: $ClusterName" (Get-ScriptName) (Get-ScriptLineNumber)
 Remove-AzureHDInsightCluster -Name $ClusterName
