+++ scripts/azure/SqlAzure/DeleteSqlAzure.ps1
@@ -4,6 +4,25 @@ Param(
     [String]$ServerName                     # required    needs to be alphanumeric
     )
 
+###########################################################
+# Start - Initialization - Invocation, Logging etc
+###########################################################
+$VerbosePreference = "SilentlyContinue"
+$ErrorActionPreference = "Stop"
+
+$scriptPath = $MyInvocation.MyCommand.Path
+$scriptDir = Split-Path $scriptPath
+
+& "$scriptDir\..\..\init.ps1"
+if(-not $?)
+{
+    throw "Initialization failure."
+    exit -9999
+}
+###########################################################
+# End - Initialization - Invocation, Logging etc
+###########################################################
+
 Write-InfoLog "Deleting SQL Server [$ServerName]" (Get-ScriptName) (Get-ScriptLineNumber)
 Remove-AzureSqlDatabaseServer $ServerName -Force
