+++ scripts/azure/Storage/DeleteStorageAccount.ps1
@@ -5,4 +5,23 @@ Param(
     [String]$AccountName                    # required    needs to be alphanumeric or "-"
     )
 
+###########################################################
+# Start - Initialization - Invocation, Logging etc
+###########################################################
+$VerbosePreference = "SilentlyContinue"
+$ErrorActionPreference = "Stop"
+
+$scriptPath = $MyInvocation.MyCommand.Path
+$scriptDir = Split-Path $scriptPath
+
+& "$scriptDir\..\..\init.ps1"
+if(-not $?)
+{
+    throw "Initialization failure."
+    exit -9999
+}
+###########################################################
+# End - Initialization - Invocation, Logging etc
+###########################################################
+
 Remove-AzureStorageAccount -StorageAccountName $AccountName
